Fixes zip so one can modify data while iterating
I was using std::make_tuple to get the type of the value in the tuple,
but I neglected to realize this would cause it to drop the reference
part of the type.  By explicitly giving it
std::tuple<iterator_deref<Container>> this is avoided.  If the iterator
derefences to a reference type, the tuple will hold a reference.