Complete rewrite of chain, safer, supports temps
Supports temporary values the same way zip does.  One thing I'd like to
be able to do is support a mix of types so something like
chain(std::vector<int>{1,2,3,4}, range(10)); would work
currently it doesnt because the iterators dereference to different types
(int& and int, respectively).  Idk if there's a safe way to handle this
that wouldn't be confusing outwardly.