Flattens chain
Chain is flattened using a similar approach to starmap.  A different
instantiation of each ChainIterWrapper for each index.  This avoids
having a nested class structure, and the time to perform operator*
doesn't increase as one gets further into the chain.  An index keeps
track of which iterator we're on each step of the way.

It's not very pretty though.