Uses constexpr array of funcion pointers for chain
Instead of the polymorphism and the base class, I was able to use a
similar approach as in starmap to get chain to act more dynamic than it
really is.  Starmap needed only one set of function pointers, for the
dereferencing, chain is admittedly more complicated because it needs one
for each operation that must be performed per iterator.