changes range to have an empty range when step=0
this seems like the best choice if I'm not gonna throw, and I'd prefer
not to throw, it seems heavyweight for the problem.  I doubt this will
be a real issue anyway. I can't personally remember ever needing to deal
with a step of 0 in python in any way, I had to go see how it behaved
when I originally wrote this range, actually.