Makes IterYield index not a reference.
There are problems with lifetime extension; binding a reference to
index won't extend the lifetime of the IterYield object (or whatever
it's called now). element isn't an obvious issue because the object it
references should live inside of Enumerable, unless the underlying
iterable's operator* returns a non-reference type in which case we're
kind of screwed.